requirements.txt
asgiref==3.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==1.3.0
Django==4.1.7
django-base64field==1.0
django-cors-headers==3.14.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.6
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.4.0